Respond to reivew comments
