+++ py/client/pydeephaven/table.py
@@ -9,14 +9,14 @@ from __future__ import annotations
 from typing import List, Union
 
 import pyarrow as pa
+
 from pydeephaven._utils import to_list
 
 from pydeephaven._table_ops import MetaTableOp, SortDirection
 from pydeephaven.agg import Aggregation
 from pydeephaven.dherror import DHError
 from pydeephaven._table_interface import TableInterface
-from pydeephaven.ticket import Ticket
-from pydeephaven.server_object import ServerObject
+from pydeephaven.ticket import Ticket, ServerObject
 from pydeephaven.updateby import UpdateByOperation
 
 
@@ -35,7 +35,7 @@ class Table(TableInterface, ServerObject):
     def table_op_handler(self, table_op):
         return self.session.table_service.grpc_table_op(self, table_op)
 
-    def __init__(self, session, ticket: Ticket, schema_header=b'', size=None, is_static=None, schema=None):
+    def __init__(self, session: 'Session', ticket: Ticket, schema_header: bytes = b'', size: int = None, is_static: bool = None, schema: pa.Schema = None):
         ServerObject.__init__(self, type="Table", ticket=ticket)
         if not session or not session.is_alive:
             raise DHError("Must be associated with a active session")
